Crate kernel64

Linux Kernel 64 Bindings

These bindings are exposed as an extended interface to the Linux kernel which are not neccesarily covered in the libc crate.

The bindings here are not SAFE in the strict sense of Rust. They require you understand the kernel interface. I will attempt to provide explainations, and links to the relevant man pages.

This bindings are ONLY valid for Linux AMD64. If you attempt to use them on other platforms this will break.

Enums

Rand
Random

Functions

copy_on_write
Copy On Write

read_only_memmap
Memory Map

Type Definitions

OSFault
So I don't have to type out io::Error all the time.