Function kernel64::copy_on_write

pub fn copy_on_write<'a, T>(buf: &'a [T]) -> Result<&'a mut [T], OSFault> where T: Sized

Copy On Write

This creates a copy on write allocation, tied to the memory lifetime of the pre-exising allocation.

COW in a nutshell allocates a new virtual memory space, that points to the old virtual memory space. When you write/modify in the new space the kernel copies the old data into a new page, and quitely updates your mappings.

Effectively this allows you to version a slice.

NOTES

Don't pass this a vector, if it resizes you can get a segfault

You WILL have to do some manual memory management.

Be Careful

mmap cares about alignment, so a too long/short array may trigger an error